# Small speedup to snpbin