@thibautjombart 2015-08-02
Looks great. Will merge after the current workshop is over, as we will rely on adegenet devel.